fix(quality): shrink new test complexity + drop redundant mockResolvedValue(undefined)
Second round of findings from the CI run after 63b8310:

1. Codacy "Complexity increased by 27 (10 max.)" on
   ui/tests/coverage-closure-regressions.test.tsx — the three new tests
   I added pushed the per-file complexity delta above Codacy's
   threshold. Collapsed the two EventsPage tests (media-query handler
   and impression-tracking analytics rejection) into a single test that
   exercises both paths in one render. Extracted mountMatchMediaMock
   (and its MatchMediaMock interface) from the test file into
   tests/page-test-helpers.tsx so the complexity lives with the other
   shared test utilities.

2. DeepSource JS-W1042 on the old coverage-closure-analytics file —
   "Remove redundant `undefined` from function call" on
   recordInteractionsSpy.mockResolvedValue(undefined). That file is
   gone; the merged-in version no longer calls mockResolvedValue with
   undefined at all (both EventsPage paths use mockRejectedValue,
   which needs an Error argument).

3. EventDetailProbe component is now a top-level named helper in the
   test file (was previously an inline IIFE-style definition inside the
   renderEventDetailProbe wrapper) which also shaved complexity.
